Replace trace ID with target
A concept of target was missing from the initial proto. A target is
required to identify where the cost is generated at. For the simplicity
of the proto, remove trace ID for now. We can always represent trace ID
in the future as an arbitrary label and we don't want to encourage
users to populate events with trace IDs for now to ensure they understand
the tradeoffs when it comes to the cardinality of events.